Change `llvm` dependency to `libllvm` for DEB
The compiler does not need the full `llvm` package. Just the runtime library `libllvm` should be enough for typical use cases. This reduces the installation size significantly: in Debian bookworm `libllvm15` installs 180 MB, while `llvm-15` installs 882 MB. Without recommends it's still 261 MB.

The full `llvm` package would only be required for building Crystal programs that link against `libllvm`, such as the compiler. So it could be cons

This change aligns with the RPM packages which depend on `llvm-libs`, and the Alpine docker image which has `libllvm` statically linked and `llvm` is not installed.